Waypoints truncated to a byte; formup banners and pins; no war NPC inside a building
WaypointList.h: getWaypointBySpecialType took the special type as a
uint8. Every war pin type (1101-1104, 9001+) was truncated before the
compare and never matched, so the engine's replace-on-add and
removeWaypointBySpecialType were silent no-ops for all of them: the map
overlay added four city pins per refresh and the earlier fix (type 0 ->
1102/1103) could not show. One word, uint8 -> int. The stock special
types are all under 256, which is why upstream never saw it.

DirectorManager: isInStructureFootprintAt(zone, x, y, margin), the
engine's structure-footprint test (what keeps a player house off a city
building) as a Lua binding; the navmesh test alone says "walkable" for
ground under a building whose footprint the mesh covers.

war_battle.lua: WarBattle.groundOk (navmesh AND not in a footprint) in
every picker -- site origins, the approach point, points along a line,
the street origin -- and a sidestep-or-refuse for a troop, a walker or a
garrison body at a point inside a building. war_orders.lua: the raid
ring too. WarSquad.attachSite carries the region.

war_squad.lua: a faction banner of the holder's side at each formup
centre (placeMarker), cleared with the areas (owner: players need to see
where the formup areas are).

war_map.lua: every staged site on the planet pinned as "Formup: <town>
site N" / "streets", one type per pin (9101 + index), in the signature.

war_probe.lua: test warSpawnSafetyCheck -- live bodies and site origins
inside a footprint, FAIL on any.

+++ MMOCoreORB/src/server/zone/objects/player/variables/WaypointList.h
@@ -105,7 +105,7 @@ public:
 		return 0;
 	}
 
-	uint64 getWaypointBySpecialType(const uint8 specialTypeID) const {
+	uint64 getWaypointBySpecialType(const int specialTypeID) const {  // SWGWar 2026-09-08: was uint8, so every type above 255 never matched (pins multiplied)
 		if (specialTypeID == 0)
 			return 0;
